# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcba54af1ef4518a5382e85
  title: Wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /tag/:
    post:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f91
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcbc504f1ef4518a5382f79
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
    get:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f99
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcbc504f1ef4518a5382f79
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
  /tag/{id}:
    delete:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f98
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcbc504f1ef4518a5382f79
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f90
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcbc504f1ef4518a5382f79
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    post:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f8e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcbc504f1ef4518a5382f79
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tag"
    
  #-----  User  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e92
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
                phone:
                  type: string
                roles:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e93
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e98
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e95
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e97
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Tag:
      x-skaffolder-id: 5dcbc504f1ef4518a5382f79
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcbc504f1ef4518a5382f77
      x-skaffolder-url: /tag
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5dcbc504f1ef4518a5382f78
          x-skaffolder-type: String
        type:
          type: string
          x-skaffolder-id-attr: 5dcbc687f1ef4518a5382f8b
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    User:
      x-skaffolder-id: 5dcba54af1ef4518a5382e8f
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcba54af1ef4518a5382e87
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8a
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8b
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e89
          x-skaffolder-type: String
          x-skaffolder-required: true
        phone:
          type: string
          x-skaffolder-id-attr: 5dcbaa9ef1ef4518a5382f19
          x-skaffolder-type: String
          x-skaffolder-unique: 
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8d
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e88
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - phone
        - username
      
    
    
  # Databases
...